Add list_data_files.pl
Supports a query based on the parsed portions of the stations
file:  country, state, station id, and station name.

Emits a list of the data files for the selection stations.

Note that a station in the stations list may not have a
data file depending on the downloaded dataset.  The effect is that
the query has an implicit "AND exists(dataset)".